@@ -26,13 +26,14 @@
     "@floating-ui/dom": "^1.6.3",
     "@fortawesome/fontawesome-free": "^6.5.1",
     "@keyv/sqlite": "^3.6.7",
-    "@skeletonlabs/skeleton": "2.9.0",
-    "@skeletonlabs/tw-plugin": "0.3.1",
+    "@skeletonlabs/skeleton": "3.0.0-next.5",
+    "@skeletonlabs/skeleton-svelte": "1.0.0-next.12",
     "@tailwindcss/forms": "0.5.7",
     "@tailwindcss/typography": "0.5.10",
-    "better-sqlite3": "^9.6.0",
+    "better-sqlite3": "^11.5.0",
     "cheerio": "^1.0.0-rc.12",
     "codemirror": "^6.0.1",
+    "cookie": "^1.0.1",
     "html-entities": "^2.5.2",
     "html-minifier": "^4.0.0",
     "keyv": "^4.5.4",
@@ -49,24 +50,24 @@
     "svelte-pdf": "^1.0.20",
     "tailwindcss": "3.4.1",
     "unified": "^11.0.4",
-    "vega": "^5.28.0",
-    "vega-embed": "^6.24.0",
-    "vega-lite": "^5.17.0",
+    "vega": "^5.30.0",
+    "vega-embed": "^6.26.0",
+    "vega-lite": "^5.21.0",
     "webdav": "^5.5.0",
     "zod": "^3.23.8"
   },
   "devDependencies": {
     "@sveltejs/adapter-auto": "^3.2.1",
     "@sveltejs/adapter-node": "^5.0.1",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@types/better-sqlite3": "^7.6.9",
     "@types/eslint": "^8.56.0",
     "@types/html-minifier": "^4.0.5",
     "@types/lodash.debounce": "^4.0.9",
     "@types/lodash.memoize": "^4.1.9",
     "@types/luxon": "^3.4.2",
-    "@types/node": "20.11.30",
+    "@types/node": "^20.11.30",
     "@types/proj4": "^2.5.5",
     "@typescript-eslint/eslint-plugin": "^7.0.0",
     "@typescript-eslint/parser": "^7.0.0",
@@ -77,15 +78,15 @@
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-drizzle": "^0.2.3",
-    "eslint-plugin-svelte": "^2.35.1",
+    "eslint-plugin-svelte": "^2.45.1",
     "postcss": "8.4.38",
     "prettier": "^3.1.1",
-    "prettier-plugin-svelte": "^3.1.2",
-    "svelte": "^4.2.7",
-    "svelte-check": "^3.6.0",
+    "prettier-plugin-svelte": "^3.2.6",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "tslib": "^2.4.1",
-    "typescript": "^5.0.0",
-    "vite": "^5.0.3",
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4",
     "vite-plugin-tailwind-purgecss": "^0.3.3",
     "vitest": "^1.6.0"
   },
