@@ -1,15 +1,19 @@
 <script lang="ts">
   import AscentTypeLabel from '$lib/components/AscentTypeLabel'
   import RouteGrade from '$lib/components/RouteGrade'
-  import type { Grade, Route, UserSettings } from '$lib/db/schema'
+  import type { Grade, UserSettings } from '$lib/db/schema'
   import type { InferResultType } from '$lib/db/types'
-  import { Ratings } from '@skeletonlabs/skeleton'
+  import { Rating } from '@skeletonlabs/skeleton-svelte'
 
   type RouteWithAscents = InferResultType<'routes', { ascents: true }>
 
-  export let route: (Omit<RouteWithAscents, 'ascents'> & Partial<Pick<RouteWithAscents, 'ascents'>>) | undefined
-  export let grades: Grade[]
-  export let gradingScale: UserSettings['gradingScale'] | undefined
+  interface Props {
+    route: (Omit<RouteWithAscents, 'ascents'> & Partial<Pick<RouteWithAscents, 'ascents'>>) | undefined
+    grades: Grade[]
+    gradingScale: UserSettings['gradingScale'] | undefined
+  }
+
+  let { route, grades, gradingScale }: Props = $props()
 
   const lastAscent = route?.ascents?.toSorted((a, b) => a.dateTime.localeCompare(b.dateTime)).at(-1)
 </script>
@@ -26,9 +30,11 @@
 
     {#if route.rating != null}
       <div>
-        <Ratings justify="start" max={3} spacing="" value={route.rating}>
-          <svelte:fragment slot="full"><i class="fa-solid fa-star text-warning-500" /></svelte:fragment>
-        </Ratings>
+        <Rating count={3} readOnly value={route.rating}>
+          {#snippet iconFull()}
+            <i class="fa-solid fa-star text-warning-500"></i>
+          {/snippet}
+        </Rating>
       </div>
     {/if}
 
