@@ -1,12 +1,16 @@
 <script lang="ts">
   import vegaEmbed, { type EmbedOptions, type VisualizationSpec } from 'vega-embed'
 
-  export let spec: VisualizationSpec
-  export let opts: EmbedOptions = {}
+  interface Props {
+    spec: VisualizationSpec
+    opts?: EmbedOptions
+  }
+
+  let { spec, opts = {} }: Props = $props()
 
   const action = (el: HTMLDivElement) => {
     vegaEmbed(el, spec, opts)
   }
 </script>
 
-<div class="vega-container w-full" use:action />
+<div class="vega-container w-full" use:action></div>
