@@ -1,5 +1,5 @@
 <script lang="ts">
-  export let data
+  let { data } = $props();
 </script>
 
 {#await import('$lib/components/BlocksMap') then BlocksMap}
