@@ -1,7 +1,7 @@
 <script lang="ts">
-  import { AppBar } from '@skeletonlabs/skeleton'
+  import { AppBar } from '@skeletonlabs/skeleton-svelte'
 
-  export let data
+  let { data } = $props()
 </script>
 
 <svelte:head>
@@ -9,18 +9,20 @@
 </svelte:head>
 
 <AppBar>
-  <svelte:fragment slot="lead">Areas</svelte:fragment>
+  {#snippet lead()}
+    Areas
+  {/snippet}
 
-  <svelte:fragment slot="trail">
+  {#snippet trail()}
     {#if data.session?.user != null}
-      <a class="btn btn-sm variant-soft-primary" href="/areas/add">
-        <i class="fa-solid fa-plus me-2" /> Add area
+      <a class="btn btn-sm preset-filled-primary-500" href="/areas/add">
+        <i class="fa-solid fa-plus"></i> Add area
       </a>
     {/if}
-  </svelte:fragment>
+  {/snippet}
 </AppBar>
 
-<div class="block card p-4 mt-8">
+<div class="block card p-4 mt-8 preset-filled-surface-100-900">
   {#if data.areas.length === 0}
     No areas yet
   {:else}
@@ -28,7 +30,7 @@
       <ul>
         {#each data.areas as area}
           <li>
-            <a href={`/areas/${area.slug}-${area.id}`}>
+            <a class="flex px-4 py-2 hover:preset-filled-primary-100-900" href={`/areas/${area.slug}-${area.id}`}>
               <span class="flex-auto">
                 {area.name}
               </span>
