@@ -3,14 +3,20 @@
   import { page } from '$app/stores'
   import AscentFormFields from '$lib/components/AscentFormFields'
   import RouteName from '$lib/components/RouteName'
-  import { AppBar, popup } from '@skeletonlabs/skeleton'
+  import { AppBar, Popover } from '@skeletonlabs/skeleton-svelte'
   import type { ActionData, PageData } from './$types'
 
-  export let data: PageData
-  export let form: ActionData
-  $: basePath = `/areas/${$page.params.slugs}/_/blocks/${$page.params.blockSlug}/routes/${$page.params.routeSlug}`
+  interface Props {
+    data: PageData
+    form: ActionData
+  }
 
-  $: grade = data.grades.find((grade) => grade.id === data.ascent.route.gradeFk)
+  let { data, form }: Props = $props()
+  let basePath = $derived(
+    `/areas/${$page.params.slugs}/_/blocks/${$page.params.blockSlug}/routes/${$page.params.routeSlug}`,
+  )
+
+  let grade = $derived(data.grades.find((grade) => grade.id === data.ascent.route.gradeFk))
 </script>
 
 <svelte:head>
@@ -24,62 +30,61 @@
 </svelte:head>
 
 <AppBar>
-  <svelte:fragment slot="lead">
+  {#snippet lead()}
     <span>Edit ascent of</span>
-    &nbsp;
     <a class="anchor" href={basePath}>
       <RouteName grades={data.grades} gradingScale={data.user?.userSettings?.gradingScale} route={data.ascent.route} />
     </a>
-  </svelte:fragment>
+  {/snippet}
 </AppBar>
 
-<form action="?/updateAscent" method="POST" use:enhance>
-  {#if form?.error != null}
-    <aside class="alert variant-filled-error mt-8">
-      <div class="alert-message">
-        <p>{form.error}</p>
-      </div>
-    </aside>
-  {/if}
+{#if form?.error}
+  <aside class="card preset-tonal-warning mt-8 p-4">
+    <p>{form.error}</p>
+  </aside>
+{/if}
 
-  <div class="mt-8">
-    <AscentFormFields
-      dateTime={form?.dateTime ?? data.ascent.dateTime}
-      filePaths={form?.filePaths ??
-        (data.ascent.files.length === 0 ? undefined : data.ascent.files.map((file) => file.path))}
-      gradeFk={form?.gradeFk ?? data.ascent.gradeFk}
-      grades={data.grades}
-      gradingScale={data.user?.userSettings?.gradingScale}
-      notes={form?.notes ?? data.ascent.notes}
-      type={form?.type ?? data.ascent.type}
-    />
-  </div>
+<form class="card mt-8 p-4 preset-filled-surface-100-900" action="?/updateAscent" method="POST" use:enhance>
+  <AscentFormFields
+    dateTime={form?.dateTime ?? data.ascent.dateTime}
+    filePaths={form?.filePaths ??
+      (data.ascent.files.length === 0 ? undefined : data.ascent.files.map((file) => file.path))}
+    gradeFk={form?.gradeFk ?? data.ascent.gradeFk}
+    grades={data.grades}
+    gradingScale={data.user?.userSettings?.gradingScale}
+    notes={form?.notes ?? data.ascent.notes}
+    type={form?.type ?? data.ascent.type}
+  />
 
   <div class="flex justify-between mt-8">
-    <button class="btn variant-ghost" on:click={() => history.back()} type="button">Cancel</button>
+    <button class="btn preset-outlined-primary-500" onclick={() => history.back()} type="button">Cancel</button>
 
     <div>
-      <button
-        class="btn variant-filled-error"
-        use:popup={{ event: 'click', target: `popup-delete-ascent`, placement: 'top' }}
-        type="button"
+      <Popover
+        arrow
+        arrowBackground="!bg-surface-200 dark:!bg-surface-800"
+        contentBase="card bg-surface-200-800 p-4 space-y-4 max-w-[320px]"
+        positioning={{ placement: 'top' }}
+        triggerBase="btn preset-filled-error-500 !text-white"
       >
-        <i class="fa-solid fa-trash me-2" />Delete ascent
-      </button>
+        {#snippet trigger()}
+          <i class="fa-solid fa-trash"></i>Delete ascent
+        {/snippet}
 
-      <button class="btn variant-filled-primary" type="submit">Save ascent</button>
+        {#snippet content()}
+          <article>
+            <p>Are you sure you want to delete this ascent?</p>
+          </article>
+
+          <footer class="flex justify-end">
+            <form method="POST" action="?/removeAscent" use:enhance>
+              <button class="btn btn-sm preset-filled-error-500 !text-white" type="submit">Yes</button>
+            </form>
+          </footer>
+        {/snippet}
+      </Popover>
+
+      <button class="btn preset-filled-primary-500" type="submit">Save ascent</button>
     </div>
   </div>
 </form>
-
-<div class="card p-4 shadow-xl" data-popup="popup-delete-ascent">
-  <p>Are you sure you want to delete this ascent?</p>
-
-  <div class="flex justify-end gap-2 mt-4">
-    <form method="POST" action="?/removeAscent" use:enhance>
-      <button class="btn btn-sm variant-filled-primary" type="submit">Yes</button>
-    </form>
-
-    <button class="btn btn-sm variant-filled-surface">Cancel</button>
-  </div>
-</div>
