@@ -3,15 +3,21 @@
   import { page } from '$app/stores'
   import AscentFormFields from '$lib/components/AscentFormFields'
   import RouteName from '$lib/components/RouteName'
-  import { AppBar } from '@skeletonlabs/skeleton'
+  import { AppBar } from '@skeletonlabs/skeleton-svelte'
   import { DateTime } from 'luxon'
   import type { ActionData, PageData } from './$types'
 
-  export let data: PageData
-  export let form: ActionData
-  $: basePath = `/areas/${$page.params.slugs}/_/blocks/${$page.params.blockSlug}/routes/${$page.params.routeSlug}`
+  interface Props {
+    data: PageData
+    form: ActionData
+  }
 
-  $: grade = data.grades.find((grade) => grade.id === data.route.gradeFk)
+  let { data, form }: Props = $props()
+  let basePath = $derived(
+    `/areas/${$page.params.slugs}/_/blocks/${$page.params.blockSlug}/routes/${$page.params.routeSlug}`,
+  )
+
+  let grade = $derived(data.grades.find((grade) => grade.id === data.route.gradeFk))
 </script>
 
 <svelte:head>
@@ -25,38 +31,33 @@
 </svelte:head>
 
 <AppBar>
-  <svelte:fragment slot="lead">
+  {#snippet lead()}
     <span>Log ascent of</span>
-    &nbsp;
     <a class="anchor" href={basePath}>
       <RouteName grades={data.grades} gradingScale={data.user?.userSettings?.gradingScale} route={data.route} />
     </a>
-  </svelte:fragment>
+  {/snippet}
 </AppBar>
 
-<form method="POST" use:enhance>
-  {#if form?.error != null}
-    <aside class="alert variant-filled-error mt-8">
-      <div class="alert-message">
-        <p>{form.error}</p>
-      </div>
-    </aside>
-  {/if}
+{#if form?.error}
+  <aside class="card preset-tonal-warning mt-8 p-4">
+    <p>{form.error}</p>
+  </aside>
+{/if}
 
-  <div class="mt-8">
-    <AscentFormFields
-      dateTime={form?.dateTime ?? DateTime.now().toSQLDate()}
-      filePaths={form?.filePaths ?? undefined}
-      gradeFk={form?.gradeFk ?? null}
-      grades={data.grades}
-      gradingScale={data.user?.userSettings?.gradingScale ?? null}
-      notes={form?.notes ?? null}
-      type={form?.type ?? null}
-    />
-  </div>
+<form class="card mt-8 p-4 preset-filled-surface-100-900" method="POST" use:enhance>
+  <AscentFormFields
+    dateTime={form?.dateTime ?? DateTime.now().toSQLDate()}
+    filePaths={form?.filePaths ?? undefined}
+    gradeFk={form?.gradeFk ?? null}
+    grades={data.grades}
+    gradingScale={data.user?.userSettings?.gradingScale ?? null}
+    notes={form?.notes ?? null}
+    type={form?.type ?? null}
+  />
 
   <div class="flex justify-between mt-8">
-    <button class="btn variant-ghost" on:click={() => history.back()} type="button">Cancel</button>
-    <button class="btn variant-filled-primary" type="submit">Save ascent</button>
+    <button class="btn preset-outlined-primary-500" onclick={() => history.back()} type="button">Cancel</button>
+    <button class="btn preset-filled-primary-500" type="submit">Save ascent</button>
   </div>
 </form>
