@@ -3,13 +3,14 @@
   import { page } from '$app/stores'
   import RouteFormFields from '$lib/components/RouteFormFields'
   import RouteName from '$lib/components/RouteName'
-  import { AppBar, popup } from '@skeletonlabs/skeleton'
+  import { AppBar, Popover } from '@skeletonlabs/skeleton-svelte'
 
-  export let data
-  export let form
-  $: basePath = `/areas/${$page.params.slugs}/_/blocks/${$page.params.blockSlug}/routes/${$page.params.routeSlug}`
+  let { data, form } = $props()
+  let basePath = $derived(
+    `/areas/${$page.params.slugs}/_/blocks/${$page.params.blockSlug}/routes/${$page.params.routeSlug}`,
+  )
 
-  $: grade = data.grades.find((grade) => grade.id === data.route.gradeFk)
+  let grade = $derived(data.grades.find((grade) => grade.id === data.route.gradeFk))
 </script>
 
 <svelte:head>
@@ -23,63 +24,62 @@
 </svelte:head>
 
 <AppBar>
-  <svelte:fragment slot="lead">
+  {#snippet lead()}
     <span>Edit route</span>
-    &nbsp;
     <a class="anchor" href={basePath}>
       <RouteName grades={data.grades} gradingScale={data.user?.userSettings?.gradingScale} route={data.route} />
     </a>
-  </svelte:fragment>
+  {/snippet}
 </AppBar>
 
-<form action="?/updateRoute" method="POST" use:enhance>
-  {#if form?.error}
-    <aside class="alert variant-filled-error mt-8">
-      <div class="alert-message">
-        <p>{form.error}</p>
-      </div>
-    </aside>
-  {/if}
+{#if form?.error}
+  <aside class="card preset-tonal-warning mt-8 p-4">
+    <p>{form.error}</p>
+  </aside>
+{/if}
 
-  <div class="mt-8">
-    <RouteFormFields
-      blockId={data.route.blockFk}
-      description={form?.description ?? data.route.description}
-      gradeFk={form?.gradeFk ?? data.route.gradeFk}
-      grades={data.grades}
-      gradingScale={data.user?.userSettings?.gradingScale}
-      name={form?.name ?? data.route.name}
-      rating={form?.rating ?? data.route.rating}
-      routeTags={form?.tags ?? data.route.tags.map((tag) => tag.tagFk)}
-      tags={data.tags}
-    />
-  </div>
+<form class="card mt-8 p-4 preset-filled-surface-100-900" action="?/updateRoute" method="POST" use:enhance>
+  <RouteFormFields
+    blockId={data.route.blockFk}
+    description={form?.description ?? data.route.description}
+    gradeFk={form?.gradeFk ?? data.route.gradeFk}
+    grades={data.grades}
+    gradingScale={data.user?.userSettings?.gradingScale}
+    name={form?.name ?? data.route.name}
+    rating={form?.rating ?? data.route.rating}
+    routeTags={form?.tags ?? data.route.tags.map((tag) => tag.tagFk)}
+    tags={data.tags}
+  />
 
-  <div class="flex justify-between mt-4">
-    <button class="btn variant-ghost" on:click={() => history.back()} type="button">Cancel</button>
+  <div class="flex justify-between mt-8">
+    <button class="btn preset-outlined-primary-500" onclick={() => history.back()} type="button">Cancel</button>
 
     <div>
-      <button
-        class="btn variant-filled-error"
-        use:popup={{ event: 'click', target: `popup-delete-route`, placement: 'top' }}
-        type="button"
+      <Popover
+        arrow
+        arrowBackground="!bg-surface-200 dark:!bg-surface-800"
+        contentBase="card bg-surface-200-800 p-4 space-y-4 max-w-[320px]"
+        positioning={{ placement: 'top' }}
+        triggerBase="btn preset-filled-error-500 !text-white"
       >
-        <i class="fa-solid fa-trash me-2" />Delete route
-      </button>
+        {#snippet trigger()}
+          <i class="fa-solid fa-trash"></i>Delete route
+        {/snippet}
 
-      <button class="btn variant-filled-primary" type="submit">Update route</button>
+        {#snippet content()}
+          <article>
+            <p>Are you sure you want to delete this route?</p>
+          </article>
+
+          <footer class="flex justify-end">
+            <form method="POST" action="?/removeRoute" use:enhance>
+              <button class="btn btn-sm preset-filled-error-500 !text-white" type="submit">Yes</button>
+            </form>
+          </footer>
+        {/snippet}
+      </Popover>
+
+      <button class="btn preset-filled-primary-500" type="submit">Update route</button>
     </div>
   </div>
 </form>
-
-<div class="card p-4 shadow-xl" data-popup="popup-delete-route">
-  <p>Are you sure you want to delete this route?</p>
-
-  <div class="flex justify-end gap-2 mt-4">
-    <form method="POST" action="?/removeRoute" use:enhance>
-      <button class="btn btn-sm variant-filled-primary" type="submit">Yes</button>
-    </form>
-
-    <button class="btn btn-sm variant-filled-surface">Cancel</button>
-  </div>
-</div>
