@@ -2,13 +2,12 @@
   import { applyAction, enhance } from '$app/forms'
   import { page } from '$app/stores'
   import RouteFormFields from '$lib/components/RouteFormFields'
-  import { AppBar, ProgressRadial } from '@skeletonlabs/skeleton'
+  import { AppBar, ProgressRing } from '@skeletonlabs/skeleton-svelte'
 
-  export let form
-  export let data
-  $: basePath = `/areas/${$page.params.slugs}/_/blocks/${$page.params.blockSlug}`
+  let { form, data } = $props()
+  let basePath = $derived(`/areas/${$page.params.slugs}/_/blocks/${$page.params.blockSlug}`)
 
-  let loading = false
+  let loading = $state(false)
 </script>
 
 <svelte:head>
@@ -16,14 +15,20 @@
 </svelte:head>
 
 <AppBar>
-  <svelte:fragment slot="lead">
+  {#snippet lead()}
     <span>Create route in</span>
-    &nbsp;
     <a class="anchor" href={basePath}>{data.block.name}</a>
-  </svelte:fragment>
+  {/snippet}
 </AppBar>
 
+{#if form?.error}
+  <aside class="card preset-tonal-warning mt-8 p-4">
+    <p>{form.error}</p>
+  </aside>
+{/if}
+
 <form
+  class="card mt-8 p-4 preset-filled-surface-100-900"
   method="POST"
   use:enhance={() => {
     loading = true
@@ -41,33 +46,25 @@
     }
   }}
 >
-  {#if form?.error != null}
-    <aside class="alert variant-filled-error mt-8">
-      <div class="alert-message">
-        <p>{form.error}</p>
-      </div>
-    </aside>
-  {/if}
-
-  <div class="mt-8">
-    <RouteFormFields
-      blockId={data.block.id}
-      description={form?.description ?? ''}
-      gradeFk={form?.gradeFk ?? null}
-      grades={data.grades}
-      gradingScale={data.user?.userSettings?.gradingScale}
-      name={form?.name ?? ''}
-      rating={form?.rating ?? null}
-      routeTags={form?.tags ?? []}
-      tags={data.tags}
-    />
-  </div>
+  <RouteFormFields
+    blockId={data.block.id}
+    description={form?.description ?? ''}
+    gradeFk={form?.gradeFk ?? null}
+    grades={data.grades}
+    gradingScale={data.user?.userSettings?.gradingScale}
+    name={form?.name ?? ''}
+    rating={form?.rating ?? null}
+    routeTags={form?.tags ?? []}
+    tags={data.tags}
+  />
 
   <div class="flex justify-between mt-8">
-    <button class="btn variant-ghost" on:click={() => history.back()} type="button">Cancel</button>
-    <button class="btn variant-filled-primary" type="submit" disabled={loading}>
+    <button class="btn preset-outlined-primary-500" onclick={() => history.back()} type="button">Cancel</button>
+    <button class="btn preset-filled-primary-500" type="submit" disabled={loading}>
       {#if loading}
-        <ProgressRadial class="me-2" width="w-4" />
+        <span class="me-2">
+          <ProgressRing size="size-4" value={null} />
+        </span>
       {/if}
       Save route
     </button>
