@@ -1,10 +1,11 @@
 <script lang="ts">
   import { page } from '$app/stores'
   import RouteName from '$lib/components/RouteName'
-  import { AppBar, Tab, TabGroup, focusTrap } from '@skeletonlabs/skeleton'
+  import { AppBar, Tabs } from '@skeletonlabs/skeleton-svelte'
   import type { Snapshot } from './$types'
+  import { onMount } from 'svelte'
 
-  export let data
+  let { data } = $props()
 
   export const snapshot: Snapshot = {
     capture: () => searchQuery,
@@ -13,13 +14,13 @@
     },
   }
 
-  let searchQuery = $page.url.searchParams.get('q') ?? ''
-  let tabSet: 'routes' | 'blocks' | 'areas' | 'users' = 'routes'
-  let element: HTMLFormElement | null = null
+  let searchQuery = $state($page.url.searchParams.get('q') ?? '')
+  let tabSet: 'routes' | 'blocks' | 'areas' | 'users' = $state('routes')
+  let element: HTMLInputElement | null = $state(null)
 
-  const formAction = (el: HTMLFormElement) => {
-    element = el
-  }
+  onMount(() => {
+    element?.focus()
+  })
 </script>
 
 <svelte:head>
@@ -27,95 +28,120 @@
 </svelte:head>
 
 <AppBar>
-  <svelte:fragment slot="lead">Search</svelte:fragment>
+  {#snippet lead()}
+    Search
+  {/snippet}
 </AppBar>
 
-<form class="mt-8" use:focusTrap={true} use:formAction>
+<form class="mt-8">
   <label class="label">
     <input
+      bind:this={element}
+      bind:value={searchQuery}
       class="input"
       name="q"
-      on:keypress={(event) => event.key === 'Enter' && element?.submit()}
+      onkeypress={(event) => event.key === 'Enter' && (element as HTMLFormElement | null)?.submit?.()}
       placeholder="Search..."
       type="search"
-      bind:value={searchQuery}
     />
   </label>
 </form>
 
 {#if data.searchResults != null}
   <div class="mt-8">
-    <TabGroup>
-      <Tab bind:group={tabSet} name="routes" value="routes">
-        <svelte:fragment slot="lead"><i class="fa-solid fa-route" /></svelte:fragment>
-        Routes ({data.searchResults.routes.length})
-      </Tab>
-      <Tab bind:group={tabSet} name="blocks" value="blocks">
-        <svelte:fragment slot="lead"><i class="fa-solid fa-mountain" /></svelte:fragment>
-        Blocks ({data.searchResults.blocks.length})
-      </Tab>
-      <Tab bind:group={tabSet} name="areas" value="areas">
-        <svelte:fragment slot="lead"><i class="fa-solid fa-layer-group" /></svelte:fragment>
-        Areas ({data.searchResults.areas.length})
-      </Tab>
-      <Tab bind:group={tabSet} name="users" value="users">
-        <svelte:fragment slot="lead"><i class="fa-solid fa-user" /></svelte:fragment>
-        Users ({data.searchResults.users.length})
-      </Tab>
+    <Tabs bind:value={tabSet}>
+      {#snippet list()}
+        <Tabs.Control value="routes">
+          {#snippet lead()}
+            <i class="fa-solid fa-route"></i>
+          {/snippet}
+          Routes ({data.searchResults.routes.length})
+        </Tabs.Control>
 
-      <svelte:fragment slot="panel">
-        {#if tabSet === 'routes'}
-          <ul class="list">
+        <Tabs.Control value="blocks">
+          {#snippet lead()}
+            <i class="fa-solid fa-mountain"></i>
+          {/snippet}
+          Blocks ({data.searchResults.blocks.length})
+        </Tabs.Control>
+
+        <Tabs.Control value="areas">
+          {#snippet lead()}
+            <i class="fa-solid fa-layer-group"></i>
+          {/snippet}
+          Areas ({data.searchResults.areas.length})
+        </Tabs.Control>
+
+        <Tabs.Control value="users">
+          {#snippet lead()}
+            <i class="fa-solid fa-user"></i>
+          {/snippet}
+          Users ({data.searchResults.users.length})
+        </Tabs.Control>
+      {/snippet}
+
+      {#snippet content()}
+        <Tabs.Panel value="routes">
+          <ul>
             {#each data.searchResults.routes as route}
-              <li class="px-4 py-2 hover:bg-primary-500/10 flex justify-between">
-                <a class="text-primary-500" href={route.pathname}>
+              <li class="hover:preset-tonal-primary flex justify-between">
+                <a class="anchor px-4 py-3 grow hover:text-white" href={route.pathname}>
                   <RouteName grades={data.grades} gradingScale={data.user?.userSettings?.gradingScale} {route} />
                 </a>
 
-                <a class="text-primary-500" href={route.block.pathname}>
+                <a class="anchor px-4 py-3 hover:text-white" href={route.block.pathname}>
                   {route.block.name}
                 </a>
               </li>
             {/each}
           </ul>
-        {:else if tabSet === 'blocks'}
+        </Tabs.Panel>
+
+        <Tabs.Panel value="blocks">
           <nav class="list-nav">
             <ul>
               {#each data.searchResults.blocks as block}
                 <li>
-                  <a href={block.pathname}>
+                  <a class="anchor px-4 py-3 flex hover:preset-tonal-primary hover:text-white" href={block.pathname}>
                     {block.name}
                   </a>
                 </li>
               {/each}
             </ul>
           </nav>
-        {:else if tabSet === 'areas'}
+        </Tabs.Panel>
+
+        <Tabs.Panel value="areas">
           <nav class="list-nav">
             <ul>
               {#each data.searchResults.areas as area}
                 <li>
-                  <a href={area.pathname}>
+                  <a class="anchor px-4 py-3 flex hover:preset-tonal-primary hover:text-white" href={area.pathname}>
                     {area.name}
                   </a>
                 </li>
               {/each}
             </ul>
           </nav>
-        {:else if tabSet === 'users'}
+        </Tabs.Panel>
+
+        <Tabs.Panel value="users">
           <nav class="list-nav">
             <ul>
               {#each data.searchResults.users as user}
                 <li>
-                  <a href={`/users/${user.userName}`}>
+                  <a
+                    class="anchor px-4 py-3 flex hover:preset-tonal-primary hover:text-white"
+                    href={`/users/${user.userName}`}
+                  >
                     {user.userName}
                   </a>
                 </li>
               {/each}
             </ul>
           </nav>
-        {/if}
-      </svelte:fragment>
-    </TabGroup>
+        </Tabs.Panel>
+      {/snippet}
+    </Tabs>
   </div>
 {/if}
