@@ -1,14 +1,15 @@
-import { join } from 'path'
-import type { Config } from 'tailwindcss'
+import { skeleton } from '@skeletonlabs/skeleton/plugin'
+import * as themes from '@skeletonlabs/skeleton/themes'
 import forms from '@tailwindcss/forms'
 import typography from '@tailwindcss/typography'
-import { skeleton } from '@skeletonlabs/tw-plugin'
+import { join } from 'path'
+import type { Config } from 'tailwindcss'
 
 export default {
   darkMode: 'class',
   content: [
     './src/**/*.{html,js,svelte,ts}',
-    join(require.resolve('@skeletonlabs/skeleton'), '../**/*.{html,js,svelte,ts}'),
+    join(require.resolve('@skeletonlabs/skeleton-svelte'), '../**/*.{html,js,svelte,ts}'),
   ],
   theme: {
     extend: {},
@@ -17,14 +18,7 @@ export default {
     forms,
     typography,
     skeleton({
-      themes: {
-        preset: [
-          {
-            name: 'rocket',
-            enhancements: true,
-          },
-        ],
-      },
+      themes: [themes.catppuccin],
     }),
   ],
 } satisfies Config
